fix test file name
